refactor: add buttons component index
