+++ src/components/buttons/XBtn.tsx
@@ -1,5 +1,5 @@
 import { X } from 'lucide-react';
 
-export default function XBtn({ onClick }: { onClick: () => void }) {
+export function XBtn({ onClick }: { onClick: () => void }) {
   return <X className="size-6 cursor-pointer" onClick={onClick} />;
 }
