+++ src/components/buttons/index.ts
@@ -1,0 +1,10 @@
+export * from './AnnouncementBtn';
+export * from './FriendAddBtn';
+export * from './GroupSettingBtn';
+export * from './LocationBtn';
+export * from './LoginBtn';
+export * from './LogoutBtn';
+export * from './MySettingBtn';
+export * from './SearchBtn';
+export * from './SetCheckBtn';
+export * from './XBtn';
